Ignore case in Record.matches
This change ensures that case is ignored when determining
whether a given `Record` matches against a specified user agent.

For example, a `Record` representing the following example:

    User-agent: Mozilla
    Disallow: /

will return `false` for `record.matches("Mozilla")`,
since the current logic lowercases the user agent passed into
the `matches` method, but does not lowercase the user agent for
the record itself.